Module 2/Order_Worker/requirements.txt
pika
mysql-connector-python
psycopg2-binary